# docs/api-doku-swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST-API for KICKERCUP REST API",
        "version": "1.0",
        "title": "KICKERCUP REST API"
    },
    "host": "kickercup-api.azurewebsites.net",
    "basePath": "/",
...
    "paths": {
...
        "/api/v1/tournaments": {
            "get": {
                "tags": [
                    "tournament-controller"
                ],
                "summary": "getAllTournaments",
                "operationId": "getAllTournamentsUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Tournament"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "tournament-controller"
                ],
                "summary": "createTournament",
                "operationId": "createTournamentUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Tournament",
                        "description": "Tournament",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tournament"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tournament"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/api/v1/tournaments/{id}": {
            "get": {
                "tags": [
                    "tournament-controller"
                ],
                "summary": "getTournamentById",
                "operationId": "getTournamentByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tournament"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "put": {
                "tags": [
                    "tournament-controller"
                ],
                "summary": "updateTournament",
                "operationId": "updateTournamentUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "TournamentDetails",
                        "description": "TournamentDetails",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tournament"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tournament"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": [
                    "tournament-controller"
                ],
                "summary": "deleteTournament",
                "operationId": "deleteTournamentUsingDELETE",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "Tournament": {
            "type": "object",
            "properties": {
                "amountGoals": {
                    "type": "integer",
                    "format": "int32"
                },
                "amountSets": {
                    "type": "integer",
                    "format": "int32"
                },
                "gamemode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "isFinished": {
                    "type": "boolean"
                },
                "isRanked": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer",
                    "format": "int64"
                }
            },
            "title": "Tournament"
        },
...
    }
}